OCI SDK 3 compatibility, add netty-based OCI http client This PR adds compatibility with OCI SDK 3.0.0. The changes for compatibility are minor. We can't upgrade to OCI SDK 3 until Micronaut 4, but this change will let users upgrade their SDK if manually.
Additionally, this PR adds an early netty-based implementation of the new OCI SDK 3 HTTP client abstraction. This avoids the dependency on jersey.

The client uses netty directly, instead of the micronaut HTTP client, because OCI is fairly particular when it comes to how the requests are structured.

In its current state, the client works. It's one of the implementations I used for testing when I designed the OCI SDK HTTP client abstraction (this is also why the tests are junit, not spock). Some features are still missing (e.g. trust store configuration), and there is still room for performance improvements (mostly connection pooling).

Fixes #434