Integrate OCI SDK metadata
This commit adds the ability to generate Micronaut metadata for OCI Java SDK
types. It does this by cloning the sources of the SDK (using the includegit
Gradle plugin), then converting the project into modules of this project.

Basically, the OCI Java SDK build is a Maven project. What we do is parsing
the various pom.xml files of this project to convert it into modules of this
project instead. The original GAV coordinates of the SDK are not kept, but
are using Micronaut GAV coordinates instead.

For each of the OCI SDK modules, we generate one synthetic project which
applies the `ocisdk-metadata-module` convention plugin. This plugin is
responsible for setting up the project as a Micronaut module, which lets
us add Micronaut annotation processors during compilation.

If a Maven module was named `bmc-waa`, with a corresponding artifact id
of `oci-java-sdk-waa`, then we create a project which name is
`oraclecloud-bmc-waa`, published at group `io.micronaut.oraclecloud`,
artifact `micronaut-oraclecloud-bmc-waa` with _this project_ version.
This artifact will contain a direct dependency to the original SDK module,
as well as a dependency on the Netty http client.

Because we don't want to actually ship binaries which contain the original
SDK classes, but only the generated ones, we then generate an additional
jar, created by the `metadataJar` task, which filters out the output of
compilation to only keep Micronaut metadata. This jar is the one which
is then published, instead of the regular jar.

Note that the converter doesn't try to be a perfect Maven build conversion:
all it cares about is being able to compile the classes, so that the
Micronaut annotation processors can run.

The result of this process can be verified by running a dummy publication.
This can be done by invoking `./gradlew pAPTBR` and then checking the
result in `build/repo`. You can then inspect the generated metadata POM
and Gradle module files, as well as the contents of the jars.

Note that doing this significantly increases configuration time for this
project, since we are now configuring around 160 modules.