Use micronaut-http-client connection pooling (#576)
This PR implements connection pooling using the existing pooling infrastructure from micronaut-http-client. There are now two providers, the old SPI NettyHttpProvider and the new ManagedNettyHttpProvider, which is managed by the ApplicationContext so that it can reuse the same DefaultHttpClient (which the user can also configure using micronaut config). AbstractSdkClientFactory requests the managed provider and injects it into managed clients.

OCI SDK config on the managed client (timeouts, SSL settings and such) does not work anymore. It should instead be configured using the `micronaut.http.services.oci` config key, like a normal http client.

The managed client still uses the same (unmanaged) OciSdkMicronautSerializer. I think it should switch to a managed serializer in the future, and I've added a todo comment for this. However it will require some more changes (jackson-databind compatibility) so I didn't do it in this PR.