#921 Allow reading body multiple times (#923)
The proposed approach to do the same as in  https://github.com/micronaut-projects/micronaut-gcp/pull/1106 did not exactly work. This is because the `InputEvent` does not allow accessing the `InputStream` directly. Instead it has a method `consumeBody(Function<InputStream>)` that should be called with implementation similar to this:
```java
<T> T consumeBody(Function<InputStream, T> function) {
      /* verify that not null and not opened twice ... */
      try {
          return function.call(this.body);
      } finally {
          this.body.close();
      }
      /* ... */
}
```
Because it closes the stream we cannot read the stream gradually and split it.

The solution is to read the stream as a whole into a byte array and then provide it to all the consumers. This will probably result in the same memory usage as in https://github.com/micronaut-projects/micronaut-gcp/pull/1106/files (since there we split each time and end up with one unused ByteBody caching the whole request). However, it will require reading the whole body before it is consumed (e.g. JSON parser cannot begin parsing the stream before whole body is read into memory).

---------

Co-authored-by: micronaut-build <65172877+micronaut-build@users.noreply.github.com>
Co-authored-by: yawkat <jonas.konrad@oracle.com>