Add shared, lazy SessionKeySupplier implementation (#940)
During startup, prof shows that a significant time is spent inside RSA key generation in the default SessionKeySupplierImpl, both in the constructor and in refreshKeys. This PR adds a new implementation that avoids generating keys until first time of use, and shares the session key between authorization providers.

Sharing between bootstrap and normal context is not yet implemented.

I have no way to test this, so can't say how much it helps.