Remove the un-required ByteBody split
As @yawkat pointed out the BaseFilterProcessor uses byteBody().split(FASTEST) to get a body for request filters and leave an extra copy for actual request parsing. Therefore, assuming that we will only parse the body once, we do not need to copy the body when it is consumed by body binder. For the case of input streams and form data we only need to parse the body once, so we do that and cache the result in a multi-values field that can be reused.

Removing the split when it is not needed would prevent from unnecessary body caching.